Add Lichtblick to GH pages (#706)
## User-Facing Changes
Users can access Lichtblick web version by only clicking the link. For
now it will be the default URL of GH pages
(https://lichtblick-suite.github.io/lichtblick/), but it can be
configured and changed to a custom one, if desired in the future.

## Description
Add pipeline to add Lichtblick web to gh-pages, the main version. It can
be triggered manually too, just in case something goes wrong, it does
not need a new release.

## Checklist

- [x] The web version was tested and it is running ok