fix: update SonarCloud workflow to include additional pull request ty… (#743)
## User-Facing Changes

N/A

## Description

This exception was added to address extreme cases when changes made by
PRs don't require a quality gate, such as code coverage, depending on
the type of change. For example, when a developer or dependabot updates
lint and code formatting libraries, where there are no logical changes,
Sonar identifies these changes as new code and requires code coverage to
be met. Therefore, by adding the "dependencies" label to the pull
request, the Sonar check continues to be performed and presented in the
pull request, but it is no longer required to proceed with the PR merge.

## Checklist

- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok
- [x] This change is covered by unit tests
- [x] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated

---------

Co-authored-by: Bezerra Luiz, (Luiz.Bezerra@ctw.bmwgroup.com) <198787532+luiz-bezerra-ctw-bmwgroup-com_QCOM@users.noreply.github.com>