Removing unsaved changes to layouts modal (#771)
## User-Facing Changes

Removing unsaved changes modal that would get not so friendly to users
because on certain panels, like 3D, a simple hover with the mouse will
trigger changes on the layout and, therefore, making the modal appear if
the user tried to switch to other layout.

**Note:** This only happened while using Lichtblick with
personal/organization layouts available.

## Description
Removed modal code and references to where it was called.

## Checklist

- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok
- [x] This change is covered by unit tests
- [x] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated