Sync main with develop (#780)
## User-Facing Changes

N/A

## Description

After a release, there's a need to synchronize the main branch with
develop, the problem is that a lot of times develop will be ahead
because not everything was released. The point of this is to
automatically create a PR when this github action is ran (it was
effectively trying to merge blindly before).
Due to the nature of github actions I forked this repo and ran a few
tests here:
https://github.com/rodrigo-rodrigues-ctw/lichtblick-test-github-actions

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

## Checklist

- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok
- [ ] This change is covered by unit tests
- [ ] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated