Making series config objects key unique for every series (#786)
## User-Facing Changes

Fixed a bug a caused by having two or more Series on a plot panel of the
same message that would cause the data to be merged into a single object
and thus creating conflicts on the line connecting the data.

<img width="826" height="481" alt="image"
src="https://github.com/user-attachments/assets/96585678-1f47-4949-aabc-e045064131c9"
/>


**Now with two series of the same message the plot renders data as
expected:** (cant really be seen here, cause as expected the data is
overlapping)


<img width="1269" height="539" alt="image"
src="https://github.com/user-attachments/assets/5544b77a-7c95-472f-8dec-e01b9f100b74"
/>


## Description

The problem lied on the moment the objects with the series configuration
were created, more specifically on their `key`. The `key` attribute
consisted of a string with the path timestamp `receiveTime` or
`headerStamp` plus the message path. So when there were two series with
the same message path the key would be the same for both of them and
therefore the objects were merged into a single one with all the data in
a single array, containing duplicated data.

<img width="1957" height="1207" alt="image"
src="https://github.com/user-attachments/assets/dba21cc3-23ca-4e66-aec1-755e75a58778"
/>

The fix I implemented is to simply include the series index on the key
which makes it impossible for it to be repeated even if they're of the
same message path.

## Checklist

- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok
- [x] This change is covered by unit tests
- [x] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated