Fix broken comparison for triangle primitive vertex update optimization (#788)
For triangle primitives there is code that checks if a new vertex/color
changed as compared to the previous state of the TriangleMesh object.
This was supposed to prevent unnecessary updates/computation.

But we noticed that the vertex comparison does not work due to internal
three.js vertex data being of type `Float32Array` whereas the primitive
vertex data is of type `Point3` with underlying `number`-values (64-bit
floating-point values). So the `vertChanged`/`colorChanged` comparisons
- even if incoming values don't change as compared to the previous step
- never evaluate to true, hence the optimization never actually works.

This PR fixes the existing optimization by rounding to the nearest
32-bit single precision float representation of both numbers before the
comparison. I added a small tolerance to the comparison for it to work
reliably. This PR uses this comparison as basis to prevent re-setting
identical vertex/color data if it hasn't changed.

---------

Signed-off-by: Thomas Sedlmayer <tsedlmayer@pmsfit.de>